Moved system message to own controller, service and repository Took the seed data out of flyway scripts and into setUp for tests
