+++ service/src/test/java/tds/config/repositories/impl/ConfigRepositoryImplIntegrationTests.java
@@ -70,114 +70,4 @@ public class ConfigRepositoryImplIntegrationTests {
 
         assertThat(result.size()).isEqualTo(0);
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
-    // ClientSystemMessage Tests
-    // -----------------------------------------------------------------------------------------------------------------
-    @Test
-    public void shouldGetOptionalEmptyClientSystemMessageForAnInvalidClientName() {
-        final String clientName = "foo";
-
-        Optional<ClientSystemMessage> result = configRepository.findClientSystemMessage(clientName, null, null, null, null, null, null);
-
-        assertThat(result.isPresent()).isFalse();
-    }
-
-    @Test
-    public void shouldGetOptionalEmptyClientSystemMessageForAnInvalidMessageKey() {
-        final String clientName = "SBAC_PT";
-        final String messageKey = "some invalid message key";
-
-        Optional<ClientSystemMessage> result = configRepository.findClientSystemMessage(clientName, messageKey, null, null, null, null, null);
-
-        assertThat(result.isPresent()).isFalse();
-    }
-
-    @Test
-    public void shouldGetEnglishLanguageClientSystemMessage() {
-        final String clientName = "SBAC_PT";
-        final String messageKey = "Your next test opportunity is not yet available.";
-        final String language = "ENU";
-        final String clientDefaultLanguage = "ENU";
-        final String context = "_CanOpenTestOpportunity";
-
-        Optional<ClientSystemMessage> result = configRepository.findClientSystemMessage(clientName, messageKey, language, clientDefaultLanguage, context, null, null);
-
-        assertThat(result.isPresent()).isTrue();
-        assertThat(result.get().getMessageId()).isEqualTo(10211);
-        assertThat(result.get().getMessage()).isEqualTo("You cannot take this test until {0}.");
-        assertThat(result.get().getLanguage()).isEqualTo(language);
-    }
-
-    @Test
-    public void shouldGetSpanishLanguageClientSystemMessage() {
-        final String clientName = "SBAC_PT";
-        final String messageKey = "Your next test opportunity is not yet available.";
-        final String language = "ESN";
-        final String clientDefaultLanguage = "ENU";
-        final String context = "_CanOpenTestOpportunity";
-
-        Optional<ClientSystemMessage> result = configRepository.findClientSystemMessage(clientName, messageKey, language, clientDefaultLanguage, context, null, null);
-
-        assertThat(result.isPresent()).isTrue();
-        assertThat(result.get().getMessageId()).isEqualTo(10211);
-        assertThat(result.get().getMessage()).isEqualTo("No puede realizar esta prueba hasta {0}.");
-        assertThat(result.get().getLanguage()).isEqualTo(language);
-    }
-
-    @Test
-    public void shouldGetDefaultLanguageClientSystemMessageWhenMissingLanguage() {
-        final String clientName = "SBAC_PT";
-        final String messageKey = "Your next test opportunity is not yet available.";
-        final String language = "NON";
-        final String clientDefaultLanguage = "ENU";
-        final String context = "_CanOpenTestOpportunity";
-
-        Optional<ClientSystemMessage> result = configRepository.findClientSystemMessage(clientName, messageKey, language, clientDefaultLanguage, context, null, null);
-
-        assertThat(result.isPresent()).isTrue();
-        assertThat(result.get().getMessageId()).isEqualTo(10211);
-        assertThat(result.get().getMessage()).isEqualTo("You cannot take this test until {0}.");
-        assertThat(result.get().getLanguage()).isEqualTo(clientDefaultLanguage);
-    }
-
-    @Test
-    public void shouldGetEnglishLanguageClientSystemMessageWhenMissingLanguageAndClientDefault() {
-        final String clientName = "SBAC_PT";
-        final String messageKey = "Your next test opportunity is not yet available.";
-        final String language = "NON";
-        final String clientDefaultLanguage = "NON";
-        final String context = "_CanOpenTestOpportunity";
-
-        Optional<ClientSystemMessage> result = configRepository.findClientSystemMessage(clientName, messageKey, language, clientDefaultLanguage, context, null, null);
-
-        assertThat(result.isPresent()).isTrue();
-        assertThat(result.get().getMessageId()).isEqualTo(10211);
-        assertThat(result.get().getMessage()).isEqualTo("You cannot take this test until {0}.");
-        assertThat(result.get().getLanguage()).isEqualTo("ENU");
-    }
-
-    // -----------------------------------------------------------------------------------------------------------------
-    // ClientLanguage Tests
-    // -----------------------------------------------------------------------------------------------------------------
-    @Test
-    public void shouldGetOptionalEmptyClientLanguageForInvalidClient() {
-        final String clientName = "invalid";
-
-        Optional<ClientLanguage> result = configRepository.findClientLanguage(clientName);
-
-        assertThat(result.isPresent()).isFalse();
-    }
-
-    @Test
-    public void shouldGetClientLanguage() {
-        final String clientName = "SBAC_PT";
-
-        Optional<ClientLanguage> result = configRepository.findClientLanguage(clientName);
-
-        assertThat(result.isPresent()).isTrue();
-        assertThat(result.get().getClientName()).isEqualTo(clientName);
-        assertThat(result.get().getDefaultLanguageCode()).isEqualTo("ENU");
-        assertThat(result.get().isInternationalize()).isTrue();
-    }
 }
