+++ service/src/test/java/tds/config/services/impl/ConfigServiceImplTest.java
@@ -79,50 +79,4 @@ public class ConfigServiceImplTest {
 
         return clientSystemFlags;
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
-    // ClientSystemMessage Tests
-    // -----------------------------------------------------------------------------------------------------------------
-    @Test
-    public void shouldGetAClientSystemMessage() {
-        String clientName = "SBAC_PT";
-        String messageKey = "some message key";
-        String languageCode = "ENU";
-        String context = "some message context";
-        String subject = "ELA";
-        String grade = "7";
-
-        ClientLanguage clientLanguage = new ClientLanguage(clientName, languageCode, true);
-        when(mockConfigRepository.findClientLanguage(clientName)).thenReturn(Optional.of(clientLanguage));
-
-        ClientSystemMessage clientSystemMessage = new ClientSystemMessage(123, "Message goes here", languageCode);
-        when(mockConfigRepository.findClientSystemMessage(clientName, messageKey, languageCode, clientLanguage.getDefaultLanguageCode(), context, subject, grade))
-            .thenReturn(Optional.of(clientSystemMessage));
-
-        String result = configService.getSystemMessage(clientName, messageKey, languageCode, context, subject, grade);
-
-        assertThat(result).isEqualTo("Message goes here [123]");
-    }
-
-    @Test
-    public void shouldGetAClientSystemMessageWithClientDefaultLanguage() {
-        String clientName = "SBAC_PT";
-        String messageKey = "some message key";
-        String languageCode = "ESN";
-        String context = "some message context";
-        String subject = "ELA";
-        String grade = "7";
-
-        ClientLanguage clientLanguage = new ClientLanguage(clientName, "ENU", false);
-        when(mockConfigRepository.findClientLanguage(clientName)).thenReturn(Optional.of(clientLanguage));
-
-        // When ClientLanguage internationalize is false, the languageCode and defaultClientLanguageCode passed in are the same since it must always use the client default
-        ClientSystemMessage clientSystemMessage = new ClientSystemMessage(123, "Message goes here", languageCode);
-        when(mockConfigRepository.findClientSystemMessage(clientName, messageKey, clientLanguage.getDefaultLanguageCode(), clientLanguage.getDefaultLanguageCode(), context, subject, grade))
-            .thenReturn(Optional.of(clientSystemMessage));
-
-        String result = configService.getSystemMessage(clientName, messageKey, languageCode, context, subject, grade);
-
-        assertThat(result).isEqualTo("Message goes here [123]");
-    }
 }
