+++ service/src/test/java/tds/config/web/endpoints/ConfigControllerTest.java
@@ -66,23 +66,4 @@ public class ConfigControllerTest {
 
         configController.getClientSystemFlag(clientName, auditObject);
     }
-
-    @Test
-    public void shouldGetClientSystemMessage() {
-        String clientName = "UNIT_TEST";
-        String messageKey = "unit test";
-        String languageCode = "Unit Language";
-        String context = "Unit context";
-        String subject = "Unit subject";
-        String grade = "Unit grade";
-        when(mockConfigService.getSystemMessage(clientName, messageKey, languageCode, context, subject, grade))
-            .thenReturn("Mocked message");
-
-        ResponseEntity<String> response = configController.getClientSystemMessage(clientName, languageCode, context, messageKey, subject, grade);
-
-        verify(mockConfigService).getSystemMessage(clientName, messageKey, languageCode, context, subject, grade);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(response.getBody()).isEqualTo("Mocked message");
-    }
 }
