+++ service/src/test/resources/db/migration/V1486282727__configs_create_tds_coremessageuser.sql
@@ -16,4 +16,3 @@ CREATE TABLE tds_coremessageuser (
   PRIMARY KEY (_fk_coremessageobject,systemid)
 );
 
-INSERT INTO tds_coremessageuser VALUES (3146, 'Database');
